+++ apps/web/contexts/index.tsx
@@ -1,6 +1,6 @@
-import { useAlertContext, AlertWrapper } from './alerts';
-import { useClientContext, ClientWrapper } from './client';
-import { useFilterContext, FilterWrapper } from './filters';
+import { useAlertContext, AlertWrapper } from './_alerts';
+import { useClientContext, ClientWrapper } from './_client';
+import { useFilterContext, FilterWrapper } from './_filters';
 
 export {
   useAlertContext,
